Wells with the component property visible="false" are rendered as plain divs.
This is a hack to provide a generic grouping option in forms. This allows groups of fields to
be conditionally revealed without having to define conditional logic on each one.

In the future we might want to consider using a custom component for this. It doesn't seem
that any of the built-in components can be easily used for only grouping, and the well was the
simplest one to provide the desired behavior.